requirements.txt
Flask==3.0.2
pypdf==4.1.0
openai==1.16.2
pyyaml==6.0.2
requests==2.32.3